backend/users-module/requirements.txt
pymongo[srv]==4.3.3
python-dotenv==1.0.0
fastapi==0.96.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
werkzeug==2.3.6
python-jose[cryptography]==3.3.0
pytest==7.4.0